fix ibc extend test
reuse start_and_wait_relayer
