+++ integration_tests/test_ibc_extended.py
@@ -1,13 +1,12 @@
 import hashlib
 import json
-import subprocess
 import time
 from pathlib import Path
 
 import pytest
-from pystarport import ports
 
-from .utils import cluster_fixture, wait_for_block, wait_for_port
+from .utils import cluster_fixture
+from .ibc_utils import start_and_wait_relayer
 
 pytestmark = pytest.mark.ibc
 
@@ -20,53 +19,6 @@ def cluster(worker_index, pytestconfig, tmp_path_factory):
         worker_index,
         tmp_path_factory.mktemp("data"),
     )
-
-
-def start_and_wait_relayer(cluster, init_relayer=True):
-    for cli in cluster.values():
-        for i in range(cli.nodes_len()):
-            wait_for_port(ports.grpc_port(cli.base_port(i)))
-
-    for cli in cluster.values():
-        # wait for at least 3 blocks, because
-        # "proof queries at height <= 2 are not supported"
-        wait_for_block(cli, 3)
-
-    # all clusters share the same root data directory
-    data_root = next(iter(cluster.values())).data_root
-    relayer = ["hermes", "--config", data_root / "relayer.toml"]
-
-    if init_relayer:
-        # create connection and channel
-        subprocess.run(
-            relayer
-            + [
-                "create",
-                "channel",
-                "--a-port",
-                "transfer",
-                "--b-port",
-                "transfer",
-                "--a-chain",
-                "ibc-0",
-                "--b-chain",
-                "ibc-1",
-                "--new-client-connection",
-                "--yes",
-            ],
-            check=True,
-        )
-
-        # start relaying
-        cluster["ibc-0"].supervisor.startProcess("relayer-demo")
-
-    rsp = json.loads(subprocess.check_output(relayer + ["query", "channels", "ibc-0"]))
-    src_channel = rsp["result"][0]["channel_id"]
-
-    rsp = json.loads(subprocess.check_output(relayer + ["query", "channels", "ibc-1"]))
-    dst_channel = rsp["result"][0]["channel_id"]
-
-    return src_channel, dst_channel
 
 
 # 3 accounts ibc test
